#283 simplify tests
